Feature - Created Auth
What
=================
Created auth routes and migrations, etc. to manage user logins and
password reset middleware/controllers, etc.

this all by running `php artisan make:auth` command.

Why
=================
To allow a user to be created in the database (and therefore called
author from now on) and to allow management of the admin area in the
future.

Ticket: #4