Enhancement - Added permalink + scopeFilter
What
=================
Added permalink database column for projects and also created the logic for its
Factory and seeder which will need to be carried across to a method for
adding records.
Added scopeFilter for when I would need to filter the posts by month and
year.

Why
=================
To be able to prefix the project sanitised name with its ID.

Ticket: #6