Bugfix - Hover elements were toggling incorrectly sometimes
What
=================
The header on single projects was toggling, regardless of whether or not
the mouse started within the hover area initially or not. So, this lead
to instances where, when the page loaded and the mouse was already
within the hover boundaries, that the title would stay within the
featured item boundaries, and then when the mouse left the title would
move aside. This is opposite to intended functionality.
The same happened for the timeline sidebar.

Issue ID: N/A